% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlanticStormsENSO-data.R
\docType{data}
\name{atlanticStormsENSO}
\alias{atlanticStormsENSO}
\title{atlanticStormsENSO data-set}
\format{a data.frame containing "year" "num_storms" and the 3-month moving average of ENSO changes}
\source{
{See the individual documentation for \code{stormReport} and \code{ENSO}, where this data-set is derived.}
}
\usage{
data(atlanticStormsENSO)
}
\description{
A data-set derived from the NHC NOAA 1950-2019 storm report and the NOAA ENSO 3-month moving average data-tracker.
The number of (hurricane) storms per year are included together with the 12 three-month moving average estimates of the ENSO changes.
}
\keyword{datasets}
