% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSO-data.R
\docType{data}
\name{ENSO}
\alias{ENSO}
\title{ENSO data-set}
\format{
a data.frame of 3-month moving averages of warming/cooling differences in the ENSO
}
\source{
{From the \href{https://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}{NHC-NOAA ENSO tracker}}
}
\usage{
data(ENSO)
}
\description{
NHC NOAA ENSO report since 1950
}
\keyword{datasets}
