% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSOscraper.R
\name{ENSOscraper}
\alias{ENSOscraper}
\title{Scrape the ENSO data-set from NHC NOAA website}
\usage{
ENSOscraper()
}
\value{
data.frame
}
\description{
{http GET call to noaa .gov for ENSO data-table. Use this to get the latest data-tables.}
}
