% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expMLE.R
\name{expMLE}
\alias{expMLE}
\title{MLE routine for exponential distributions}
\usage{
expMLE(x)
}
\arguments{
\item{x}{the sample of random-variates of waiting times to use in the MLE routine}
}
\value{
fitDistr object
}
\description{
{Wrapper to \code{fitdist} from \code{fitdistrplus} for exponential distribution distributions}
}
