% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterStormType.R
\name{filterStormType}
\alias{filterStormType}
\title{Filter storm data-set by tropical storm or hurricane}
\usage{
filterStormType(dta, exclude = TRUE, type = "Tropical")
}
\arguments{
\item{dta}{the data-set of all storms, from \code{stormReportScraper()}}

\item{exclude}{boolean: whether to exclude by keyword or include}

\item{type}{either "Tropical" or "Hurricane" or "All"}
}
\value{
data.frame
}
\description{
{Filter the data-set by tropical storms or hurricanes}
}
\details{
{Note early data-points do not record storm names by whether they are tropical storms or hurricanes.}
}
