% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStormData.R
\name{loadStormData}
\alias{loadStormData}
\title{Wrapper to load in storm data quickly}
\usage{
loadStormData(exclude = TRUE, type = "Tropical")
}
\arguments{
\item{exclude}{whether to exclude data by keyword or include}

\item{type}{"Tropical" "Hurricane" or "All"}
}
\value{
list
}
\description{
{Use \code{exclude = TRUE} and "Tropical" to get back named storms + hurricanes.
Use \code{exclude = FALSE} and "Hurricanes" to get back only hurricanes, etc.}
}
