% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStormENSO.R
\name{loadStormENSO}
\alias{loadStormENSO}
\title{Load a data-frame of number of storms in the atlantic basin plus ENSO moving averages}
\usage{
loadStormENSO(exclude = TRUE, type = "Tropical")
}
\arguments{
\item{exclude}{whether to exclude by keyword type or only include}

\item{type}{the keyword either "All", "Tropical" or "Hurricane"}
}
\value{
data.frame
}
\description{
{Load a data.frame of number of hurricanes and ENSO moving averages per year.
You can exclude all storms with "Tropical Storm" in their name by setting \code{exclude = TRUE} and \code{type = "Tropical"}, etc. Pacific
storms are missing reports prior to the 90s so they cannot be used here.}
}
