% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCI.R
\name{meanCI}
\alias{meanCI}
\title{Compute sample mean and confidence intervals}
\usage{
meanCI(x, alpha = 0.05)
}
\arguments{
\item{x}{random variates to compute mean of}

\item{alpha}{significance level, defaults to 0.05}
}
\value{
data.frame
}
\description{
{Computes the sample mean and confidence interval around the estimate}
}
\details{
{For samples \eqn{\leq 45} in size, the Student-t distribution is used to compute the CI, otherwise the normal distribution is used.}
}
