% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonFit.R
\name{poissonFit}
\alias{poissonFit}
\title{Fit a Poisson distribution to a set of count data of some phenomena over a time-period}
\usage{
poissonFit(countsData, alpha = 0.05, verbose = TRUE)
}
\arguments{
\item{countsData}{the vector of number of events per time period}

\item{alpha}{the significance level}

\item{verbose}{(boolean) whether or not to print output on function call}
}
\value{
list
}
\description{
{This function runs a Chi-square test and fits the MLE Poisson parameter estimate to a given data-set of counts over time.}
}
\details{
{A wrapper to \code{chiSquareTest} and \code{poissonMLE}}
}
