% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonMLE.R
\name{poissonMLE}
\alias{poissonMLE}
\title{MLE routine for Poisson distributions}
\usage{
poissonMLE(x)
}
\arguments{
\item{x}{the sample of random-variates to use in the MLE routine}
}
\value{
fitDistr object
}
\description{
{Wrapper to \code{fitdist} from \code{fitdistrplus} for Poisson distributions}
}
