% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootingCounts.R
\name{shootingCounts}
\alias{shootingCounts}
\title{Return number of mass-shootings per month from the Mother-jones data-set}
\usage{
shootingCounts()
}
\value{
table
}
\description{
Return number of mass-shootings per month from the Mother-jones data-set
}
