% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootingDemographics.R
\name{shootingDemographics}
\alias{shootingDemographics}
\title{Return demographics of mass-shooters and frequency of occurence by state}
\usage{
shootingDemographics(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean whether to print on call or not}
}
\value{
list
}
\description{
Return demographics of mass-shooters and frequency of occurence by state
}
