% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shootingTimes.R
\name{shootingTimes}
\alias{shootingTimes}
\title{Compute waiting times between mass shootings in the USA using the mother-jones data-set}
\usage{
shootingTimes()
}
\value{
data.frame
}
\description{
Compute waiting times between mass shootings in the USA using the mother-jones data-set
}
