% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormReport-data.R
\docType{data}
\name{stormReport}
\alias{stormReport}
\title{stormReport data-set}
\format{a list of all storm reports with year, basin and name}
\source{
{From the \href{https://www.nhc.noaa.gov/TCR_StormReportsIndex.xml}{NHC-NOAA storm report}}
}
\usage{
data(stormReport)
}
\description{
NHC NOAA storm report since 1958
}
\keyword{datasets}
