% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormReportScraper.R
\name{stormReportScraper}
\alias{stormReportScraper}
\title{Download 1954-present Storm Report from NHC NOAA .gov}
\usage{
stormReportScraper()
}
\value{
data.frame containing \code{year}, \code{basin} and \code{name} of the storm.
}
\description{
{Get the 1954-present Storm report from NHC NOAA .gov}
}
\details{
{This function makes a GET call to the NHC NOAA .gov site, where they make available in xml format the entire 1954-present Storm Report.
The data is read into a data.frame and ordered by year.}
}
